Updating spinner examples
